Addressing cran comments. comment: -------- The description text seems to contain too many blank spaces. Please check that.
response:
--------
I've removed the extra spaces.

comment:
--------
If there are references describing the methods in your package, please add these in the description field of your DESCRIPTION file in the form
authors (year) <doi:...>
authors (year) <arXiv:...>
authors (year, ISBN:...)
or if those are not available: <https:...>
with no space after 'doi:', 'arXiv:', 'https:' and angle brackets for auto-linking.
(If you want to add a title as well please put it in quotes: "Title")

response:
--------
This package contains a shiny module for interfacing with command-line tools and does not contain any unique methods to cite.

comment:
--------
Please make sure that you do not change the user's options, par or working directory. If you really have to do so within functions, please ensure with an *immediate* call of on.exit() that the settings are reset when the function is exited. e.g.:
...
oldwd <- getwd()           # code line i
on.exit(setwd(oldwd))        # code line i+1
...
setwd(...)            # somewhere after
...
e.g.: R/NCA_Server.R
If you're not familiar with the function, please check ?on.exit. This function makes it possible to restore options before exiting a function even if the function breaks. Therefore it needs to be called immediately after the option change within a function.

response:
--------
Oops :). Sorry I missed that. I've fixed it.